joined simclr em_comm_ssl game
