--- egg/zoo/emcom_as_rl/simclr_replica/sweeps/simclr.json
@@ -1,11 +1,0 @@
-{
-    "n_epoch": [100],
-    "batch_size": [128],
-    "lr": [2.4],
-    "random_seed": [111],
-    "output_size": [2048],
-    "fp16": [true],
-    "validation_freq": [1],
-    "validation_dataset_dir": ["/datasets01/imagenet_full_size/061417/val"],
-    "wandb": [true]
-}
