+++ egg/zoo/emcom_as_ssl/communication/images/small_set_imagenet.dat
