+++ egg/zoo/emcom_as_ssl/communication/scripts/get_img_paths.py
