+++ egg/zoo/emcom_as_ssl/sweeps/fixed_temp_non_shared_vision.json
