+++ egg/zoo/emcom_as_ssl/sweeps/fixed_temp_w_decay_shared_vision_bn.json
